+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Add search box to Search & filter pane [UILISTS-252]
 * Add filter by List Creator and Updated by [UILISTS-253]
+* Filter lists by source (canned/system generated list) [UILISTS-263]
+* Sort list of lists by name and updated date [UILISTS-254]
 
 ## [5.0.2](https://github.com/folio-org/ui-lists/tree/v5.0.2) (2026-05-04)